UCP/DEVICE/API: Review comments.

+++ src/ucp/api/device/ucp_device_types.h
@@ -47,7 +47,7 @@ typedef struct ucp_device_mem_list_handle {
     /**
      * Number of entries in the memory descriptors array @a elems.
      */
-    size_t          mem_list_length;
+    uint32_t        mem_list_length;
 
     /**
      * Array of pointers to UCT device endpoints, used for multi-lane
@@ -63,22 +63,22 @@ typedef struct ucp_device_mem_list_handle {
     /**
      * Array of local addresses for the device transfer operations.
      */
-    void     **local_addrs;
+    void            **local_addrs;
 
     /**
      * Array of remote addresses for the device transfer operations.
      */
-    uint64_t *remote_addrs;
-    
+    uint64_t        *remote_addrs;
+
     /**
      * Array of lengths of the local buffers in bytes.
      */
-    size_t   *lengths;
+    size_t          *lengths;
 
     /**
      * Array of UCT memory element objects.
      */
-    void *uct_mem_elements;
+    void            *uct_mem_elements;
 
     /**
      * local address, remote address, and length arrays, are allocated contiguously.
