+++ src/ucp/api/device/ucp_host.h
@@ -162,7 +162,8 @@ ucp_device_mem_list_create(ucp_ep_h ep,
  *
  * @param [in] handle     Descriptors list handle.
  */
-size_t ucp_device_get_mem_list_length(const ucp_device_mem_list_handle_h handle);
+uint32_t
+ucp_device_get_mem_list_length(const ucp_device_mem_list_handle_h handle);
 
 
 /**
